styled desktop, nav info button
