+++ client/scss/main.scss
@@ -24,7 +24,6 @@ $orange-dark: #c45d45;
 /* Import Assets */
 @import './nav.scss';
 @import './commandPane.scss';
-@import './about.scss';
 @import './board.scss';
 @import './commandQueue.scss';
 @import './runButton.scss';
